Fix floating point tests in clockwork

+++ apps/hardware_benchmarks/tests/fp_conv_1_1/fp_conv_1_1_generator.cpp
@@ -31,7 +31,7 @@ public:
         conv(x, y)  += fp_kernel(r.x, r.y) * hw_input(x + r.x, y + r.y);
 
         Func hw_output("hw_output");
-        hw_output(x, y) = f32(conv(x, y));
+        hw_output(x, y) = u8(conv(x, y));
         output(x, y) = u8(ceil(hw_output(x,y)));
 
         /* THE SCHEDULE */
