+++ apps/hardware_benchmarks/tests/fp_conv_6_6/fp_conv_6_6_generator.cpp
@@ -37,7 +37,7 @@ public:
         conv(x, y)  += fp_kernel(r.x + 6* r.y) * hw_input(x + r.x, y + r.y);
 
         Func hw_output("hw_output");
-        hw_output(x, y) = f32(conv(x, y));
+        hw_output(x, y) = u8(conv(x, y));
         output(x, y) = u8(ceil(hw_output(x,y)));
         /* THE SCHEDULE */
         if (get_target().has_feature(Target::CoreIR)) {
